backend/agent/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
pydantic-settings==2.2.1
langgraph==0.1.9
langchain==0.2.1
langchain-community==0.2.1
langchain-openai==0.1.8
langchain-anthropic==0.1.11
httpx==0.27.0
redis==5.0.4
prometheus-fastapi-instrumentator==6.1.0
python-dotenv==1.0.1
